create initial repository documentation
We need documentation that describes how contributers and users would
use our new open source project - end users that want to install it or
developers that want to contribute

README file with high level overview and instructions on next steps -
this is the landing page for anyone looking at the project. So likely
woould need a lot of iterations on this.

CONTRIBUTING.md which describes expected contribution process for people
who want to participate in the project

Type of change:
[x] Bug fix/small improvement