versatile-data-kit: Include link to public Slack workspace
The main README file now features a link to the Slack workspace
in the Contributing section. Additionally, a TODO comment is
included because we need to set up a slackin web app to make
joining the workspace easier, as currently it can only happen
by invite, which is not ideal.

Testing done: I clicked on the link

Signed-off-by: gageorgiev <gageorgiev@vmware.com>