vdk plugins: set README.md as setup long description
Long description option of setup.py is showned as description in the
PyPY UI as Project Description - similarly to how it is for
https://test.pypi.org/project/vdk-control-cli

But our plugins do not set long descriptoin field so it looks like this
: https://test.pypi.org/project/vdk-trino/0.1.347213486/ (empty).

This is not very friendly nor useful to end users.

Added to setup.py of all plugins to set long descriptoin their README.md
file. This is pretty standard practice for this

Testing Done:  For all of the plugins installed them locally using `pip
install -e .` and verified long description is set using `python
setup.py --long-description`
Also uploaded one of the plugins manually to test.pipy.org: E.g:
https://test.pypi.org/project/vdk-test-utils/0.1.dev1

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>