vdk-test-utils: improve README.md
As our project is becoming public, we need better documentation
explaining what problems each plugin solves and how to use it.

Improved this README by adding better overview and Usage section. I
think it'd be good to iterate a few more times in next few weeks/months
on the README.