vdk-plugin-control-cli: change long_description_content_type to markdown
Upon upload to pipy.org it expects that README.md is correctly formatted
restructuredtext format. And we were not
-https://gitlab.com/vmware-analytics/versatile-data-kit/-/jobs/1477157238

But this can be overriden. And we override it.

Alternative approach I considered was adding a hook to validate synxtax
of markdown files to be correct restructuredtext format:

I tried using https://github.com/Lucas-C/pre-commit-hooks-markup but it
didn't work it validated the file ok but twine upload still failed.
I guess they are not pypi compatible as they say or I did something
wrong.

I tried also https://github.com/myint/rstcheck but this was even worse
it was a lot of more spammy but did not catch the issue pypi linters
catches.

Testing Done: twine upload of it successfully.