vdk-core: Fix vdk-core CICD triggering on unrelated changes
vdk-core CICD jobs were triggered as a result of changes made to
the control-service part of the repository. This should not happen,
so this change fixes that by making sure that the simple-functional-test
and the plugin build jobs only trigger when there are changes to the
vdk-core part of the repository.

Additionally, plugin build jobs will now only trigger on changes to the
specific plugin's directory, not on any change to the main vdk-core
directory.

Signed-off-by: gageorgiev <gageorgiev@vmware.com>