gitlab-runners: start runners in cicd namespace
Until now they were started in default namespace.
For troubleshooting purposes it's better if they are visible in the
common namespace for now. If we start having too many things in the
namespace we can split them.

Testing Done: ran install-runners.sh script locally and saw runners are
started in cicd namespace

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>