vdk-control-cli:enable relase acceptance test
Enable release acceptance test. We have a CICD deployment that is
working.

Control Service is deployed under http://cicd-control-service-svc:8092,
This is only internal URI and available only in kubernetes

Outside of Kubernetes it can be access to
http://a5424653167db4d01b3acf66ec2ac8d9-481011980.us-west-1.elb.amazonaws.com:8092/swagger-ui.html

Testing Done: this pull request CI (I have temporarily enabled to run
acceptance test on PR. Before merging will set it to run on main only.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>