vdk-trino: Add fact snapshot template
Templates give users an easy way to perform some common SQL
transformations. We try to implement them as atomic, idempotent and
optimized operations. Also, their implementation can later be changed
and improved without the need to change anything on the user side.

Add fact snapshot template which is used for Periodic Snapshot Fact
Tables data loading strategy.

Unit tests added for the template - one for the common case with valid
source and target tables, one for the case where the source table is
empty.

The commit is a non-breaking change which adds new functionality.

Signed-off-by: Yana Zhivkova (yzhivkova@vmware.com)