vdk-trino: Implement SCD2 template without alter table rename
Trino "alter table rename" operation is quite expensive
in certain cases (Hive/S3), since it requires copying the whole table
to a new location.
For this reason, the option is disabled by default for
production tables. For this reason, we add an option to
use templates without relying on table rename.

In order to preserve the "rename table" option, we introduce
a config var which defines what strategy will be used when
writing data to target table -
TRINO_TEMPLATES_DATA_TO_TARGET_STRATEGY. Possible values
for now are - "RENAME" and "INSERT_SELECT".
The newly introduced INSERT_SELECT strategy functionality
creates the target table, inserts data in it from source
and drops source.

New unit tests are added for testing different strategies.
Also, template tests are called for both possible strategies.

The commit is a non-breaking change which adds functionality.

Signed-off-by: Yana Zhivkova (yzhivkova@vmware.com)