vdk-heartbeat: Introduce VDK Heartbeat
As part of the development of the Versatile Data
Kit project, we need to make sure all components
work as expected. This necessitates the creation
of a heartbeat test to be performed before merging
any changes to the versatile data kit components.

This change introduces the vdk heartbeat test,
which aims at performing last check of the vdk sdk,
and make sure there are no regressions in terms of
functionality that end-users would expect.

Testing Done: unit tests

Signed-off-by: Andon Andonov <andonova@vmware.com>