control service cicd: add deploy demo/cicd environment scripts
As part of our CICD we are starting full deployment of Control Service.
It is upgraded automatically to use latest version of Control Service
and latest version of vdk (we would use quickstart-vdk).

It's used to run end to end acceptance teast (vdk-heartebat) against it
to verify all happy paths from user point of view are working correctly.

It's used sometimes for demo purposes to demonstrate the API for
example.

Testing Done: ran the script locally and deployed it.
Control Service can be seen at
http://a5424653167db4d01b3acf66ec2ac8d9-481011980.us-west-1.elb.amazonaws.com:8092/swagger-ui.html
. It requires authorization for any operation.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>