vdk-heartbeat: Fix vdk-heartbeat CI/CD script
As part of the Continuous Delivery process of
Versatile Data Kit, we need to ensure all CI/CD
scripts work as expected.

This change fixes a bug where the name of the main
branch in the CI/CD script was specified as `master`,
instead of `main`.

Testing Done: CI/CD run

Signed-off-by: Andon Andonov <andonova@vmware.com>