control-service: Create Github repository on Helm install
As part of the effort to improve the experience of users
willing to try out the control service, we are making
changes to the Helm chart to allow installation with
minimal effort.

This commit makes Git parameters of the Helm installation optional.

If no Git parameters are specified, the Helm installation will
automatically create a new public Git repository in
https://github.com/versatile-data-kit-jobs. The name of the
new repository can be specified via the `deploymentGitRepoName`
parameter. If no name is specified, the name will be a randomly
generated UUID.

Finally, if specified, the `deploymentGitUrl` will take precedence
over `deploymentGitRepoName`.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>