control-service: Update Helm chart
In order to simplify installation of Helm chart some features are
disabled by default.

The serviceMonitor and alerting are disabled by default.
WARNING: When this change is adopted, the respective values must
be switched back to true: `metrics.serviceMonitor.enabled` and
`alerting.enabled'.

The `SPRING_SECURITY_OAUTH2_RESOURCESERVER_JWT_JWK_SET_URI` is not
set if the respective jwtJwkSetUri property is empty, to prevent
spring security from throwing an exception.

Clean up the README file. Update the NOTES.txt file to indicate
that the installation will use a public Git repository.

The changes are tested by manually installing the Helm chart to a
Minikube cluster and verifying that the installation and the service
start are successful.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>