versatile-data-kit: Include a non-expiring Slack invite link
The previous plan of using a slackin web app to streamline
Slack invitations fell through due to it being reliant on
legacy Slack software, and not working anymore. An alternative
approach of using a non-expiring invite link was discovered and
implemented.

Testing done: used the link through a different PC and
successfully joined the Slack workspace.

Signed-off-by: gageorgiev <gageorgiev@vmware.com>