quickstart-vdk: release docker image with bundled vdk
In order to be used by Control Service runtime environment a
distribution or packaging of vdk need to also release a docker image.

We are doing this for quickstart-vdk here.

Testing Done: Gitlab CI - temporarily changed release stage to trigger
on pull requests in order to verify image is built correctly.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>