vdk-trino: Fact snapshot template unit tests for backup restore/fail
Fact snapshot template might lose its target table in its last step.
In order to make the template idempotent and to prevent losing data,
we create a backup table. The first step of the template tries to
recover target (if it's missing) from this backup.

Add tests for successfully and unsuccessfully recovering target
from back.

Templates unit tests.

The commit is a non-breaking change which adds new unit tests.

Signed-off-by: Yana Zhivkova (yzhivkova@vmware.com)