control-service: publish job-builder image
Job Builder Image is used by Control Service to build the data job image
when a new data job is being deployed. It's started in kubernetes job -
this way the operation can be done asyncrounously and also enables
certain level of extensibility as the job-builder image can be
configured.

We need to publish the image to dockerhub on each change . This is
enabling the publish step and setting correct variables to use.

Testing Done executed locally ./publish-vdk-job-builder.sh , tested in
CI by starting step in PR (temporarily)