vdk-trino: Address snapshot (append_overwrite) template feedback
Since the template is not a classic snapshot implementation, its
documentation and name need to be changed. There is no need for the
backup table's name to be a parameter.

Remove incorrect README.md, it will be revised and added in a different
commit. Add TODO for renaming the template.
Add method for getting backup table name.

Templates unit tests.

The commit is a non-breaking change which does minor refactoring.

Signed-off-by: Yana Zhivkova (yzhivkova@vmware.com)