vdk-ingest-http: ingest any json data
The vdk-ingest-http plugin provides simple
mechanism to ingest data over http.

The change removes unnecessary check and
refactors the exception messages shown in
the logs. The removed check is not necessary
as the plugin should not impose strict
restrictions on the data format. In its
current form, the plugin only needs to make
sure that either a destination_table value
is provided by the user, or that a @table key
is present in the payload. The correct value
for the destination table is responsibility
of the user.

Testing Done: Added unit tests, and
refactored old ones.

Signed-off-by: Andon Andonov <andonova@vmware.com>