vdk-trino: Automatically start Trino server for templates local testing
Trino templates require a running Trino server. Until now,
a server had to be manually started, so that they could pass locally.

A plugin is used for starting the server automatically - pytest-docker.
Fixtures are implemented in conftest.py and Trino tests use them.
trino_service fixture waits for the server to be fully up and running by checking if a query can be run against it.

Unit tests and PR CICD.

Signed-off-by: Yana Zhivkova <yzhivkova@vmware.com>