vdk-trino: Rename template from snapshot to periodic_snapshot
We want to make the template name as self-explanatory as
possible. The current snapshot template implements the
Periodic Snapshot Fact Table strategy, but there is also
the Accumulating Snapshot Fact Table strategy.

All occurrences of the word snapshot in the template implementation and
tests (both in files content and file names) changed to
periodic_snapshot.

Tested by unit tests.

Signed-off-by: Yana Zhivkova <yzhivkova@vmware.com>