versatile-data-kit: Include DL join link
After this change, members of the VMware organization
can join the DL by clicking on the link provided in the
contact section. Unfortunately, we couldn't find a trivial
way to automate the join process for non-members, so a new
group was established which external users can email to
request to join. This can probably be automated through
some script (see the python-dev DL for inspiration),
but this is not a priority as we don't expect there to
be an overwhelming surge of users looking to join our
mailing list.

Signed-off-by: gageorgiev <gageorgiev@vmware.com>