control-service: expose datajob base image in helm (#112)
We use data job base image to specify the base image used when
generating data job specific image. It's used to install common
utilities and native libraries. Until now it was not directly configured
in the helm chart mostly because the default was good enough for current
users. But making it accessible to helm would:

a) make it easier for users to realize it's possible to customize data
job image. The current default customization was created in away that
works for our current users but we cannot extend it forever. It's just
an example.

b) by putting in helm it would allow users to use docker proxy cache as
registry - the current default in dockerhub would hit dockerhut rate
limits and they cannot override the registry only easily.

Testing Done: cd to helm chart and helm install foo . --dry-run and
inspected the deployment yaml generated the image is correctly set.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>