vdk-core: check correct ingestion format
With the current ingestion mechanism, versatile
data kit is intended to support only json serializable
data payloads. However, currently it does not
properly handle payloads that are not json-compatible.

This change fixes the issue, and adds check that
ensures that no corrupt payloads would be accepted
for ingestion.

Testing Done: Updated unit tests

Signed-off-by: Andon Andonov <andonova@vmware.com>