vdk-control-cli: Team in execute start output
When a data job is manually executed using the Execution
API, a message is outputted with an `execute --show`
command that the client can use to see the status of the
execution. The command in the message, however, was not
complete, as it did not show the team name.

This change fixes that by adding the team name to the
show command in the output message.

Testing Done: Manually tested to see if the team name
appears in the output message of `vdkcli execute --start`.
It does.
Additional unit tests added

Signed-off-by: Andon Andonov <andonova@vmware.com>