vdk-sqlite: Fix directory issue (#123)
Until now, vdk-sqlite would expect a directory
as a variable, where it would expect to find a
db file called vdk-sqlite.db. This meant that
users cannot connect to a preexisting database
with a filename different from 'vdk-sqlite.db',
or they would have to rename the file, which is
clunky and annoying. After this change, you can
set the database file directly, and subsequently
can have it named in whatsoever way.

Testing done: unit test

Signed-off-by: gageorgiev <gageorgiev@vmware.com>