[draft]vkd-heartbeat: Add execution API to vdk heartbeat (#116)
Recently we added a new execution api to the control service.
This change aims to include the new api functionality in the
vdk-heartbeat tests.

This change  adds the new functionality of the execute API to the
VDK heartbeat tests. We now have functionality to
check if a data job has a running data jobexecution
and start a data job execution through job_controller.py

The new test flow disables the data job deployment after all other
tests have passed and then checks for a runnig execution.
When there are none we start a new data job execution through the
execution API, and checks that the data job execution has started. Since
the deployment was disabled the only data job execution
that could start is a manual one through the execution api.
Once that execution finishes the test concludes and execution
resources get cleaned up as normal.

Testing: CI/CD

Signed-off-by: Momchil Zhivkov <mzhivkov@vmware.com>