vdk-control-cli: improve execute error handling
When an exception occurs while the `vdkcli execute`
command is used, we print the whole stack trace in the
terminal, which is not user-friendly, and could cause
confusion.

This change adds error handling to the execute command
that provides more manageable error message, and avoids
printing the whole stack trace of the exception.

Testing Done: Existing unit tests and added a new
test

Signed-off-by: Andon Andonov <andonova@vmware.com>