control-service: do not log stracktrace in debug log
When MetricsHelper failed to serialize (for telemetry) some object we
log strack trace , which is distracting when troubleshooting issues. As
telemetry is informational it's possible that some object cannot be
serialized to json in which case we just send its string representation
which is ok.

Testing Done: run Control service locally and no longer saw stack trace
with the debug message

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>