control-pane: Add vdkcli execute --cancel
Why:
This change is a part of our new Data Job Execution API.
Currently we have the functionality to start and query
running executions. We want to add the ability to also
cancel already running executions.

What:
Introduced changes are in control-plane, since we already
had the API endpoint functional, however control plane would
throw an exception. To cancel a running data job from a cron
job this change uses the K8S api's deleteNamespacedJob method.

This change is a part of the Data Job Execution Api in Production
epic.

Testing:
Added tests.

Signed-off-by: Momchil Zhivkov <mzhivkov@vmware.com>