vdk-control-cli: Add debug logging for errors
For debugging purposes we need to be able to log
exception stacktraces, even when we handle the
exceptions.

This change adds debug log to the ApiClientErrorDecorator,
in order to log stacktraces, exception classes and
method names. To view these logs, --dev mode or --verbosity
set to DEBUG is required when using vdkcli.

Testing Done: Run a vdkcli command locally with verbosity
set to DEBUG, and examined that the stacktrace is present

Signed-off-by: Andon Andonov <andonova@vmware.com>