vdk-core: allow iterable for send_tabular_data_for_ingestion
We currently check if rows object in send_tabular_data_for_ingestion is
iterator. But that is unnecessarily restrictive.

Iterable is a parent class of Iterator and enables traversing the
collection in for cycle.
See https://docs.python.org/3/library/collections.abc.html

Also, there could be a plugin that could infer the column names from
their order so requiring column names always is not necessary. That's
often seen in practice so we probably should allow it.

Testing Done: existing unit tests

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>