control-service: Introduce Executions in GraphQL Schema/API
Why:
Job executions are important part of the data jobs details. They could be easily attached to the existing graphql schema/endpoint in order to extend the capabilities for the consumers.

What is done:
GraphQL schema is changed in order to provide future functionality for paging, sorting and filtering with the native graphql-way. Same mechanism with querying data jobs is applied, but this time as arguments in the executions fields

Testing:
Local run of tests and manual testing

Type of change:
Improvement of the current GraphQL Schema

Signed-off-by: Plamen Kostov <pkostov@vmware.com>