vdk-core: Make ingest method router store method keys in lowercase
Before this change, in the case where we've registered
an ingester with the key 'sqlite', but the env variable
VDK_INGEST_METHOD_DEFAULT is set to 'SQLITE', the job
will fail with an error stating that the expected
ingest method was not found.
It seems reasonable that if these two values are the
same, ignoring case, then ingestion should work as
normally.
This change addresses this by amending the ingester
router to store the key of the ingester in lowercase,
and lowercase the method string before initializing
the ingester.

Testing done: locally tested by testing out different
combinations of casing for the values for the key
stated in the plugin initialization function andthe
env variable VDK_INGEST_METHOD_DEFAULT.

Signed-off-by: gageorgiev <gageorgiev@vmware.com>