control-service: Enable sending logs to log insight.
Why: Recently we have had problems with deployment of data jobs.
Currently the control service only sends logs to the container's
stdout stream. This meant that deployment errors could be lost
due to log rolling. We have decided as a team to start sending
logs to log insight. This introduces a change that enables us
to do so.

What:Added a logback-spring.xml file which contains the properties
for logging. By default we will send logs to stdout still. However,
we now check if a particular environment variable is present and if
it is we also send logs to the host provided in the environment var.

Testing: Tested starting the service locally with and without the
specified env variable. Logs appear in loginsight when variable is
present. In both cases logs appear in stdout.

Signed-off-by: Momchil Zhivkov <mzhivkov@vmware.com>