control-service: Add username to executions
As part of improving the user experience when
doing manual job executions, and troubleshooting
failed executions, we need to be able to tell by
whom was an execution started.

This change adds the username of the user, who has
initiated an execution, to the startedBy annotation.

Testing Done: Updated existing unit test to reflect the
change.

Signed-off-by: Andon Andonov <andonova@vmware.com>