vdk-sqlite: Ensure ingestion uses the implemented SQLite connection
Until now, ingestion to SQLite would use a database
connection established by sqlalchemy's create_engine
module. This was subomptimal, due to the fact that
the vdk-sqlite plugin already offers a connection
module, and in the future we might want to implement
some particular functionality unique to the vdk-sqlite
plugin.

After this change, ingestion now uses the connection
native to the SQLite plugin.

Testing done: ingested to local database, unit tests

Signed-off-by: gageorgiev <gageorgiev@vmware.com>