vdk-core: Fix ingestion wrong destination_table
The Versatile Data Kit needs to allow for data to be
ingested in different destination tables within the same
job run.

This change fixes a bug, where data was ingested only in
one destination table, instead of all respective destination
tables. The change is in IngestionBase and tracks changes to
destination table based on payloads passed.

Testing Done: Unit tests passed, and tested locally by
running a data job and verifying that the data was
ingested in the correct destination tables

Signed-off-by: Andon Andonov <andonova@vmware.com>