vdk-csv: add new plugin - vdk ingest-csv command (#161)
* vdk-csv: add new plugin - vdk ingest-csv command

Ingesting a CSV file (or excel or other tabular file that can be easily
converted to csv)  into a database so that it can be analysized is a
very frequent use-case. A lot of tools allow one-off
export of data into csv which can be ingested one-off. Also excel or
similar is used to store manually edited data which people may ingest
for analysis.

We want to make this as easy as possible for users to do so.

We expose command vdk ingest-csv which user can point to csv file and we
automatically will ingesting using Ingestion API so they can all
benefitsfrom it - users can specify configruable destination, the
ingestion is done in parallel with support for batching requests and so
on.

It allows to specify table name and extra options.
We need table name so that users can customize the destination table. We
infer it from file name if not specified.

Extra options enable users to handle corner cases in ingesting csv file.
Those extra options are same as pandas.read_csv. Our contract for
reading csv file is the same as pandas.read_csv .
And we enable the same customization as pandas.read_csv.

Testing Done: added functional test and tested manually
```
pip install vdk-ingest-file
export VDK_INGEST_METHOD_DEFAULT=file
vdk ingest-csv -f test.csv
```
And verified that file test.xxx.json is created with "ingested" data.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>