vdk-core: Add test for ingestion to tables
As part of a bux fix, a change to the IngesterBase
class, to make possible ingesting data to multiple destination
tables, was made without adding a test to ensure everything
works as expected.

This change adds a new unit test, which tests if sending
payloads to different destination tables works as expected.

Testing Done: Added new unit test for testing
ingestion to multiple destination tables

Signed-off-by: Andon Andonov <andonova@vmware.com>