vdk-core: fix misconfigured description breaks help
As description is is passed by 3th party plugins and can easily be
misconfigured - for example passing wrong type. We are converting it to
string to make sure it's the correct type and can be handled.

It's better than raising an error probably as 1 meaningless description
should not break the whole plugin

Testing Done: unit tests

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>