vdk-core: fix issue in default values overriding set values (#172)
ConfigurationBuilder is passed to vdk_configure hook so that plugins can
add different configurations. Ideas is to enable plugins to boht define
their conifguraiton keys they need and to enable other plugins to
populate those keys. When adding new key users can also set a default
value (if no value is set by other plugin).

Pluggy enables to configure hook order - set hook implementation with
tryFirst=True will be first tryLast=True it will be last and if neither
is set it will be in the middle.
So it's possible that a plugin which adds a new configuration key
defintion to be executed after. This causes an issue where default value
override legimate value.

This attempts to fix that.

We should consider if having two hooks - vdk_configure_define and
vdk_configure_set might be better ? It adds yet one more hook to
maintain but it may make it less error prone ot use. But it requires
bigger discussion so it's out of scope for this change.

Testing Done: added unit tests to cover the issue

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>