vdk-core: make sure ingest failure causes data job failure (#173)
It's very very very important that data job fails if we fail to
upload/ingest/send all data to target destination. As it gives immediate
feedback to user (either directly when running locally or through VDK
monitoring and notifications when running on cloud) that something has
went wrong and make sure they know that not all data has been ingested.

If we swollow the errors or do not report the failure and the data job
report success, users will not find out (even if we log warnings and
errors). Job is sucessful and users would not have a reason to look more
carefully in the logs. And of course when running on cloud as mentioned
above , nor monitoring nor notifications will be triggered.

Testing Done: tests, locally ingested data using vdk ingest-csv

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>