vdk-sqlite: Ensure target is captured reasonably when ingesting
Before this change, vdk-sqlite would rely on the ingestion router
to send it the correct target. However, the router does not validate
the value of target, so it was possible to ingest to None as a base.

After this change, vdk-sqlite's ingester checks the supplied target,
the VDK_INGEST_TARGET_DEFAULT and the VDK_SQLITE_FILE for the value
of target in that order, taking the first non-None value, and raises
an exception if all of them are None.

Testing done: pipelines passed

Signed-off-by: gageorgiev <gageorgiev@vmware.com>