vdk-server: fix git server installation
The initial implementation of the Git server installation had some flaws
that are resolved as part of this commit.

* The server configuration could fail if the server has not started yet
even though the container was running;
* The configuration request appears to fail because it automatically
redirects to /user/login after installation;
* The request for the Git repository creation was failing because the
parameters were not passed in correctly;

Testing done: locally running the plugin and verifying that it was running
and the Git server was created and configured properly and the repository
was created.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>