vdk-server: Wire up external containers (Git and Docker) with the service
Updated the Helm installation to account for the previously installed Git
and Docker containers. This includes obtaining and passing the IP of the
Git container directly because the Job Builder fails to resolve it.

Reduced the number of replicas of the control service and the database to 1.

Added missing dependencies to setup.py.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>