vdk-server: expose control service outside of the cluster
Make the service visible outside of the cluster by installing an
ingress nginx controller. The Ingress k8s object itself is installed
via Helm chart because the current version of the kubernetes Python
client has issues installing Ingress objects.

Testing done: manually perform the installation and verify that the
ingress is correctly installed and that the service is accessible.
The url of the service is 'localhost' (without port or protocol), so
an example command looks like:

```
vdk list --all -u localhost
```

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>