control-service: bugfix on the execution cancellation api
why: During testing, when making calls to the endpoint I noticed
that the job execution state would get overriden by the control
service. The cancel functionality would execute and set the status
of the execution to cancelled, but then it would get overriden by the
JobExecution watcher functionality. Also as a user if I try to cancel
an execution in running state which is no longer present in K8S and the
API returns 404 (job not found) we would return an error and the execution
would still be in running state.

what: Added a check in the update method which checks if the execution is
in cancelled state and if it is, don't update it. Cancelled is set only when
the delete call on K8S has finished successfully. Added another check in the
K8S call to check if the API returns 404. If it does, we mark the execution
as cancelled in the database.

type: Bugfix

testing: CI/CD, ran tests locally

Signed-off-by: Momchil Zhivkov <mzhivkov@vmware.com>