control-service: clean up Git repository creation in the Helm chart
Since we've moved away from the idea to provision a public Github
repository during Helm chart installation, revert the previous
changes to the Helm chart.

Testing done: Installed the chart manually in a local Kind cluster
and verified that the installation is successful and no git repo
is created during the process.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>