vdk-core: Remove reference to taurus from main namespace
Until now, the main namespace of the vdk-core project was called
taurus. This is not ideal, as taurus is the name of an internal
VMware project, while Versatile Data Kit will soon be open source.

This change fixes that by renaming the taurus namespace to vdk,
and the sub-namespace which was previously called vdk to internal.

Testing done: pipelines passed

Signed-off-by: gageorgiev <gageorgiev@vmware.com>