vdk-control-cli: make vdk create smarter (#211)
* vdk-control-cli: make vdk create smarter

We need to be able to use vdk without control service. The normal
workflow we recommend is

See
https://github.com/vmware/versatile-data-kit/wiki/User-Guide#data-jobs-development-workflow

Basically:
1. User would create job with vdk create
2. develop their job and run it locally with vdk run
3. vdk deploy ...

It's only on step 3 that control service is necessary so neither step 1
or 2 should fail if it is not avaialble.

But currently step 1 fails without control service.

This change fixes that by making vdk create with no arguments a a bit
smarter- it would behave as local if control service is not configured
otherwise it will try to register the job

Testing Done: unit tests including new ones. Also tried the flow
manually vdk create and was promted for name and team and got a sample
local job.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>