vdk-snowflake: Refactor plugin code
After receiving feedback on the code structure of the
vdk-snowflake plugin, a refactoring was necessary to
optimize it.

This change optimizes the configuration of the plugin
and changes the result message printed after executing
a query to use tabular format, instead of json string.

Testing Done: Added unit test and tested locally by
running a data job and queries

Signed-off-by: Andon Andonov <andonova@vmware.com>