vdk-core: Fix ingesting NULL values
Until now, None values wouldn't be ingested and would just be
skipped. This is not preferred, as it's possible for a database
to have NULL values within it.
This change fixed that by removing the skipping of None values
and instead converting them to the string "NULL".

Testing done: pipelines passed

Signed-off-by: gageorgiev <gageorgiev@vmware.com>