vdk-core: add termination message writer plugin (#219)
In order to complete the alerting and Execution API integrations we need
to export the job execution status. The status will be exported through
the K8S Pod termination message. Control Service monitors it and records it
into the database.

This plugin writes the job execution status to a specific file (default: "/dev/termination-log").
https://kubernetes.io/docs/tasks/debug-application-cluster/determine-reason-pod-failure/#writing-and-reading-a-termination-message

Testing Done: added unit tests

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com