vdk-control-cli: Make sample job be runnable without error
The sample job created by `vdk create` was previously not runnable
due to the steps executed - they expect the target database to have
a table called `hello_world`. This change amends the steps to still
describe the functionality offered by Data Jobs but to now be
runnable.
Additionally, this change makes small amendments to the docs of the
Data Job.

Testing done: executed `vdk create`, then ran the job and saw it
succeed.

Signed-off-by: gageorgiev <gageorgiev@vmware.com>