versatile-data-kit: Fix --help command truncation (#225)
The command help messages printed when executing `vdk --help`
were being truncated. This change fixes that by trimming the size of
the first sentence in the help message, or including a short_help
parameter in the command decorator. Additionally, it capitalizes
all occurences of "Data Job" in the command's help text.

Testing done: executed `vdk --help` and checked the messages

Signed-off-by: gageorgiev <gageorgiev@vmware.com>