vdk-plugin-control-cli: Remove vdkcli from help/err msgs
vdk-plugin-control-cli is dependent on
vdk-control-cli and uses its commands but
the console application of vdk-core is called vdk,
not vdkcli. We need vdk's user-facing documentation to refer
to vdk not vdkcli.

Replace all occurrences of vdkcli in help and error
messages with vdk. Add comment about vdkcli/vdk difference
in README.md.

Checked "vdk <command> --help" for all commands.
Also, I called them with "vdk" and they seem to work.

Signed-off-by: Yana Zhivkova <yzhivkova@vmware.com>