vdk-server: configure vdk to use local Control Service during installation
As part of the local installation of the Control Service, we want the
Control Service to be immediatelly useable with vdk cli commands without
the need for the user to specify the REST endpoint for each command.

Extended the set-default command to also support rest-api-url in addition
to team. Changed the vdk server command to use the set-default option in
order to configure default REST API url.

Testing done: Manually performed a Control Service installation by using
`vdk server -i` then verified that all vdk cli commands work without the
need to specify the --rest-api-url option.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>